nookLibrary: Made generation of screen saver of covers properly asynchronous.
The generation was accidentally not actually done in the background, thus potentially resulting in "force-close" prompts to the user.  Additionally, instead of an AsyncTask a regular Thread is now used because of issues Android 1.5 has executing multiple instances of the same AsyncTask instance (any execution of a new instance after one instance has been run would execute onPreExecute() but not doInBackground()).

Furthermore, the code recently added by Hari down-scaling cover images in order to avoid out-of-memory conditions with the gallery has been adapted to keep the original image along with the scaled-down version (as "FILE.orig.EXTENSION") because the "screen saver of covers" is rendered rather moot without high-resolution covers.  (The reason why we create a differently-named original version rather than a differently-named small version is that ScannedFile#setCover() is called in various places and occasionally with the result of ScannedFile#getCover() as argument, which will change the "main" cover file as registered internally to the small version (and re-run the down-scaling checks on it).  This is also cleaner in that ScannedFile#setCover() won't silently change its internal notion while the caller may continue referring to the cover file originally passed.)

Also removed a special-case optimization in the screen saver generation code, the reason being that I overlooked some further conditions rendering the optimization even more of a corner case and thus unwarranted.